requirements-dev.txt
amqp==2.6.1
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==4.4.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
flower==0.9.7
humanize==3.5.0
iniconfig==1.1.1
kombu==4.6.11
packaging==20.9
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.18
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
vine==1.3.0
wcwidth==0.2.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dj-stripe==2.5.1
python-decouple==3.5
django-cors-headers==3.10.0
whitenoise==5.3.0
Pillow==8.4.0
django-taggit==2.0.0
django-filter==21.1